rm core.jl; create core_functions.jl, core_structures.jl; fix #39
